requirements.txt
Flask==2.0.2
flask-cors==3.1.1
numpy==1.21.3
Pillow==9.0.0
torch==1.10.0
opencv-python==4.5.4.66
